+++ examples/sqlite/db_sqlite.nim
@@ -22,8 +22,8 @@
 ## Parameter substitution
 ## ----------------------
 ##
-## All ``db_*`` modules support the same form of parameter substitution.
-## That is, using the ``?`` (question mark) to signify the place where a
+## All `db_*` modules support the same form of parameter substitution.
+## That is, using the `?` (question mark) to signify the place where a
 ## value should be placed. For example:
 ##
 ## .. code-block:: Nim
@@ -35,7 +35,7 @@
 ##
 ## .. code-block:: Nim
 ##
-##    import db_sqlite
+##    import std/db_sqlite
 ##
 ##    # user, password, database name can be empty.
 ##    # These params are not used on db_sqlite module.
@@ -66,7 +66,7 @@
 ##
 ## .. code-block:: nim
 ##
-##    import db_sqlite, math
+##    import std/[db_sqlite, math]
 ##
 ##    let db = open("mytest.db", "", "", "")
 ##
@@ -94,16 +94,67 @@
 ##
 ##    db.close()
 ##
+## Storing binary data example
+##----------------------------
+##
+## .. code-block:: nim
+##
+##   import std/random
+##
+##   ## Generate random float datas
+##   var orig = newSeq[float64](150)
+##   randomize()
+##   for x in orig.mitems:
+##     x = rand(1.0)/10.0
+##
+##   let db = open("mysqlite.db", "", "", "")
+##   block: ## Create database
+##     ## Binary datas needs to be of type BLOB in SQLite
+##     let createTableStr = sql"""CREATE TABLE test(
+##       id INTEGER NOT NULL PRIMARY KEY,
+##       data BLOB
+##     )
+##     """
+##     db.exec(createTableStr)
+##
+##   block: ## Insert data
+##     var id = 1
+##     ## Data needs to be converted to seq[byte] to be interpreted as binary by bindParams
+##     var dbuf = newSeq[byte](orig.len*sizeof(float64))
+##     copyMem(unsafeAddr(dbuf[0]), unsafeAddr(orig[0]), dbuf.len)
+##
+##     ## Use prepared statement to insert binary data into database
+##     var insertStmt = db.prepare("INSERT INTO test (id, data) VALUES (?, ?)")
+##     insertStmt.bindParams(id, dbuf)
+##     let bres = db.tryExec(insertStmt)
+##     ## Check insert
+##     doAssert(bres)
+##     # Destroy statement
+##     finalize(insertStmt)
+##
+##   block: ## Use getValue to select data
+##     var dataTest = db.getValue(sql"SELECT data FROM test WHERE id = ?", 1)
+##     ## Calculate sequence size from buffer size
+##     let seqSize = int(dataTest.len*sizeof(byte)/sizeof(float64))
+##     ## Copy binary string data in dataTest into a seq
+##     var res: seq[float64] = newSeq[float64](seqSize)
+##     copyMem(unsafeAddr(res[0]), addr(dataTest[0]), dataTest.len)
+##
+##     ## Check datas obtained is identical
+##     doAssert res == orig
+##
+##   db.close()
+##
 ##
 ## Note
 ## ====
 ## This module does not implement any ORM features such as mapping the types from the schema.
-## Instead, a ``seq[string]`` is returned for each row.
+## Instead, a `seq[string]` is returned for each row.
 ##
 ## The reasoning is as follows:
-## 1. it's close to what many DBs offer natively (char**)
+## 1. it's close to what many DBs offer natively (`char**`:c:)
 ## 2. it hides the number of types that the DB supports
-## (int? int64? decimal up to 10 places? geo coords?)
+##    (int? int64? decimal up to 10 places? geo coords?)
 ## 3. it's convenient when all you do is to forward the data to somewhere else (echo, log, put the data into a new query)
 ##
 ## See also
@@ -120,7 +171,7 @@ import sqlite3, macros
 import db_common
 export db_common
 
-import std/private/since
+import std/private/[since, dbutils]
 
 type
   DbConn* = PSqlite3  ## Encapsulates a database connection.
@@ -160,24 +211,16 @@ proc dbQuote*(s: string): string =
   add(result, '\'')
 
 proc dbFormat(formatstr: SqlQuery, args: varargs[string]): string =
-  result = ""
-  var a = 0
-  for c in items(string(formatstr)):
-    if c == '?':
-      add(result, dbQuote(args[a]))
-      inc(a)
-    else:
-      add(result, c)
+  dbFormatImpl(formatstr, dbQuote, args)
 
 proc prepare*(db: DbConn; q: string): SqlPrepared {.since: (1, 3).} =
-  ## Creates a new ``SqlPrepared`` statement.
+  ## Creates a new `SqlPrepared` statement.
   if prepare_v2(db, q, q.len.cint,result.PStmt, nil) != SQLITE_OK:
     discard finalize(result.PStmt)
     dbError(db)
 
 proc tryExec*(db: DbConn, query: SqlQuery,
-              args: varargs[string, `$`]): bool {.
-              tags: [ReadDbEffect, WriteDbEffect].} =
+              args: varargs[string, `$`]): bool =
   ## Tries to execute the query and returns `true` if successful, `false` otherwise.
   ##
   ## **Examples:**
@@ -199,8 +242,7 @@ proc tryExec*(db: DbConn, query: SqlQuery,
       discard finalize(stmt)
       result = false
 
-proc tryExec*(db: DbConn, stmtName: SqlPrepared): bool {.
-              tags: [ReadDbEffect, WriteDbEffect].} =
+proc tryExec*(db: DbConn, stmtName: SqlPrepared): bool =
     let x = step(stmtName.PStmt)
     if x in {SQLITE_DONE, SQLITE_ROW}:
       result = true
@@ -208,8 +250,7 @@ proc tryExec*(db: DbConn, stmtName: SqlPrepared): bool {.
       discard finalize(stmtName.PStmt)
       result = false
 
-proc exec*(db: DbConn, query: SqlQuery, args: varargs[string, `$`])  {.
-  tags: [ReadDbEffect, WriteDbEffect].} =
+proc exec*(db: DbConn, query: SqlQuery, args: varargs[string, `$`]) =
   ## Executes the query and raises a `DbError` exception if not successful.
   ##
   ## **Examples:**
@@ -242,21 +283,25 @@ proc setupQuery(db: DbConn, stmtName: SqlPrepared): SqlPrepared {.since: (1, 3).
 
 proc setRow(stmt: PStmt, r: var Row, cols: cint) =
   for col in 0'i32..cols-1:
-    setLen(r[col], column_bytes(stmt, col)) # set capacity
-    setLen(r[col], 0)
-    let x = column_text(stmt, col)
-    if not isNil(x): add(r[col], x)
+    let cb = column_bytes(stmt, col)
+    setLen(r[col], cb) # set capacity
+    if column_type(stmt, col) == SQLITE_BLOB:
+      copyMem(addr(r[col][0]), column_blob(stmt, col), cb)
+    else:
+      setLen(r[col], 0)
+      let x = column_text(stmt, col)
+      if not isNil(x): add(r[col], x)
 
 iterator fastRows*(db: DbConn, query: SqlQuery,
-                   args: varargs[string, `$`]): Row {.tags: [ReadDbEffect].} =
+                   args: varargs[string, `$`]): Row =
   ## Executes the query and iterates over the result dataset.
   ##
   ## This is very fast, but potentially dangerous. Use this iterator only
   ## if you require **ALL** the rows.
   ##
   ## **Note:** Breaking the `fastRows()` iterator during a loop will cause the
-  ## next database query to raise a `DbError` exception ``unable to close due
-  ## to ...``.
+  ## next database query to raise a `DbError` exception `unable to close due
+  ## to ...`.
   ##
   ## **Examples:**
   ##
@@ -288,8 +333,7 @@ iterator fastRows*(db: DbConn, query: SqlQuery,
   finally:
     if finalize(stmt) != SQLITE_OK: dbError(db)
 
-iterator fastRows*(db: DbConn, stmtName: SqlPrepared): Row
-                  {.tags: [ReadDbEffect,WriteDbEffect], since: (1, 3).} =
+iterator fastRows*(db: DbConn, stmtName: SqlPrepared): Row {.since: (1, 3).} =
   discard setupQuery(db, stmtName)
   var L = (column_count(stmtName.PStmt))
   var result = newRow(L)
@@ -301,8 +345,7 @@ iterator fastRows*(db: DbConn, stmtName: SqlPrepared): Row
     dbError(db)
 
 iterator instantRows*(db: DbConn, query: SqlQuery,
-                      args: varargs[string, `$`]): InstantRow
-                      {.tags: [ReadDbEffect].} =
+                      args: varargs[string, `$`]): InstantRow =
   ## Similar to `fastRows iterator <#fastRows.i,DbConn,SqlQuery,varargs[string,]>`_
   ## but returns a handle that can be used to get column text
   ## on demand using `[]`. Returned handle is valid only within the iterator body.
@@ -340,8 +383,7 @@ iterator instantRows*(db: DbConn, query: SqlQuery,
   finally:
     if finalize(stmt) != SQLITE_OK: dbError(db)
 
-iterator instantRows*(db: DbConn, stmtName: SqlPrepared): InstantRow
-                      {.tags: [ReadDbEffect,WriteDbEffect], since: (1, 3).} =
+iterator instantRows*(db: DbConn, stmtName: SqlPrepared): InstantRow {.since: (1, 3).} =
   var stmt = setupQuery(db, stmtName).PStmt
   try:
     while step(stmt) == SQLITE_ROW:
@@ -372,8 +414,7 @@ proc setColumns(columns: var DbColumns; x: PStmt) =
     columns[i].tableName = $column_table_name(x, i)
 
 iterator instantRows*(db: DbConn; columns: var DbColumns; query: SqlQuery,
-                      args: varargs[string, `$`]): InstantRow
-                      {.tags: [ReadDbEffect].} =
+                      args: varargs[string, `$`]): InstantRow =
   ## Similar to `instantRows iterator <#instantRows.i,DbConn,SqlQuery,varargs[string,]>`_,
   ## but sets information about columns to `columns`.
   ##
@@ -434,7 +475,7 @@ proc len*(row: InstantRow): int32 {.inline.} =
   column_count(row)
 
 proc getRow*(db: DbConn, query: SqlQuery,
-             args: varargs[string, `$`]): Row {.tags: [ReadDbEffect].} =
+             args: varargs[string, `$`]): Row =
   ## Retrieves a single row. If the query doesn't return any rows, this proc
   ## will return a `Row` with empty strings for each column.
   ##
@@ -470,7 +511,7 @@ proc getRow*(db: DbConn, query: SqlQuery,
   if finalize(stmt) != SQLITE_OK: dbError(db)
 
 proc getAllRows*(db: DbConn, query: SqlQuery,
-                 args: varargs[string, `$`]): seq[Row] {.tags: [ReadDbEffect].} =
+                 args: varargs[string, `$`]): seq[Row] =
   ## Executes the query and returns the whole result dataset.
   ##
   ## **Examples:**
@@ -491,14 +532,13 @@ proc getAllRows*(db: DbConn, query: SqlQuery,
   for r in fastRows(db, query, args):
     result.add(r)
 
-proc getAllRows*(db: DbConn, stmtName: SqlPrepared): seq[Row]
-                {.tags: [ReadDbEffect,WriteDbEffect], since: (1, 3).} =
+proc getAllRows*(db: DbConn, stmtName: SqlPrepared): seq[Row] {.since: (1, 3).} =
   result = @[]
   for r in fastRows(db, stmtName):
     result.add(r)
 
 iterator rows*(db: DbConn, query: SqlQuery,
-               args: varargs[string, `$`]): Row {.tags: [ReadDbEffect].} =
+               args: varargs[string, `$`]): Row =
   ## Similar to `fastRows iterator <#fastRows.i,DbConn,SqlQuery,varargs[string,]>`_,
   ## but slower and safe.
   ##
@@ -524,12 +564,11 @@ iterator rows*(db: DbConn, query: SqlQuery,
   ##    db.close()
   for r in fastRows(db, query, args): yield r
 
-iterator rows*(db: DbConn, stmtName: SqlPrepared): Row
-              {.tags: [ReadDbEffect,WriteDbEffect], since: (1, 3).} =
+iterator rows*(db: DbConn, stmtName: SqlPrepared): Row {.since: (1, 3).} =
   for r in fastRows(db, stmtName): yield r
 
 proc getValue*(db: DbConn, query: SqlQuery,
-               args: varargs[string, `$`]): string {.tags: [ReadDbEffect].} =
+               args: varargs[string, `$`]): string =
   ## Executes the query and returns the first column of the first row of the
   ## result dataset. Returns `""` if the dataset contains no rows or the database
   ## value is `NULL`.
@@ -558,28 +597,34 @@ proc getValue*(db: DbConn, query: SqlQuery,
     if cb == 0:
       result = ""
     else:
-      result = newStringOfCap(cb)
-      add(result, column_text(stmt, 0))
+      if column_type(stmt, 0) == SQLITE_BLOB:
+        result.setLen(cb)
+        copyMem(addr(result[0]), column_blob(stmt, 0), cb)
+      else:
+        result = newStringOfCap(cb)
+        add(result, column_text(stmt, 0))
   else:
     result = ""
   if finalize(stmt) != SQLITE_OK: dbError(db)
 
-proc getValue*(db: DbConn,  stmtName: SqlPrepared): string
-              {.tags: [ReadDbEffect,WriteDbEffect], since: (1, 3).} =
+proc getValue*(db: DbConn,  stmtName: SqlPrepared): string {.since: (1, 3).} =
   var stmt = setupQuery(db, stmtName).PStmt
   if step(stmt) == SQLITE_ROW:
     let cb = column_bytes(stmt, 0)
     if cb == 0:
       result = ""
     else:
-      result = newStringOfCap(cb)
-      add(result, column_text(stmt, 0))
+      if column_type(stmt, 0) == SQLITE_BLOB:
+        result.setLen(cb)
+        copyMem(addr(result[0]), column_blob(stmt, 0), cb)
+      else:
+        result = newStringOfCap(cb)
+        add(result, column_text(stmt, 0))
   else:
     result = ""
 
 proc tryInsertID*(db: DbConn, query: SqlQuery,
-                  args: varargs[string, `$`]): int64
-                  {.tags: [WriteDbEffect], raises: [].} =
+                  args: varargs[string, `$`]): int64 =
   ## Executes the query (typically "INSERT") and returns the
   ## generated ID for the row or -1 in case of an error.
   ##
@@ -606,13 +651,13 @@ proc tryInsertID*(db: DbConn, query: SqlQuery,
     discard finalize(stmt)
 
 proc insertID*(db: DbConn, query: SqlQuery,
-               args: varargs[string, `$`]): int64 {.tags: [WriteDbEffect].} =
+               args: varargs[string, `$`]): int64 =
   ## Executes the query (typically "INSERT") and returns the
   ## generated ID for the row.
   ##
   ## Raises a `DbError` exception when failed to insert row.
-  ## For Postgre this adds ``RETURNING id`` to the query, so it only works
-  ## if your primary key is named ``id``.
+  ## For Postgre this adds `RETURNING id` to the query, so it only works
+  ## if your primary key is named `id`.
   ##
   ## **Examples:**
   ##
@@ -635,21 +680,18 @@ proc insertID*(db: DbConn, query: SqlQuery,
   if result < 0: dbError(db)
 
 proc tryInsert*(db: DbConn, query: SqlQuery, pkName: string,
-                args: varargs[string, `$`]): int64
-               {.tags: [WriteDbEffect], raises: [], since: (1, 3).} =
+                args: varargs[string, `$`]): int64 {.since: (1, 3).} =
   ## same as tryInsertID
   tryInsertID(db, query, args)
 
 proc insert*(db: DbConn, query: SqlQuery, pkName: string,
-             args: varargs[string, `$`]): int64
-            {.tags: [WriteDbEffect], since: (1, 3).} =
+             args: varargs[string, `$`]): int64 {.since: (1, 3).} =
   ## same as insertId
   result = tryInsert(db, query,pkName, args)
   if result < 0: dbError(db)
 
 proc execAffectedRows*(db: DbConn, query: SqlQuery,
-                       args: varargs[string, `$`]): int64 {.
-                       tags: [ReadDbEffect, WriteDbEffect].} =
+                       args: varargs[string, `$`]): int64 =
   ## Executes the query (typically "UPDATE") and returns the
   ## number of affected rows.
   ##
@@ -672,11 +714,11 @@ proc execAffectedRows*(db: DbConn, query: SqlQuery,
   result = changes(db)
 
 proc execAffectedRows*(db: DbConn, stmtName: SqlPrepared): int64
-                      {.tags: [ReadDbEffect, WriteDbEffect],since: (1, 3).} =
+                      {.since: (1, 3).} =
   exec(db, stmtName)
   result = changes(db)
 
-proc close*(db: DbConn) {.tags: [DbEffect].} =
+proc close*(db: DbConn) =
   ## Closes the database connection.
   ##
   ## **Examples:**
@@ -687,12 +729,11 @@ proc close*(db: DbConn) {.tags: [DbEffect].} =
   ##    db.close()
   if sqlite3.close(db) != SQLITE_OK: dbError(db)
 
-proc open*(connection, user, password, database: string): DbConn {.
-  tags: [DbEffect].} =
+proc open*(connection, user, password, database: string): DbConn =
   ## Opens a database connection. Raises a `DbError` exception if the connection
   ## could not be established.
   ##
-  ## **Note:** Only the ``connection`` parameter is used for ``sqlite``.
+  ## **Note:** Only the `connection` parameter is used for `sqlite`.
   ##
   ## **Examples:**
   ##
@@ -711,8 +752,7 @@ proc open*(connection, user, password, database: string): DbConn {.
   else:
     dbError(db)
 
-proc setEncoding*(connection: DbConn, encoding: string): bool {.
-  tags: [DbEffect].} =
+proc setEncoding*(connection: DbConn, encoding: string): bool =
   ## Sets the encoding of a database connection, returns `true` for
   ## success, `false` for failure.
   ##
@@ -792,7 +832,7 @@ macro untypedLen(args: varargs[untyped]): int =
 
 template exec*(db: DbConn, stmtName: SqlPrepared,
           args: varargs[typed]): untyped =
-  when args.untypedLen > 0:
+  when untypedLen(args) > 0:
     if reset(stmtName.PStmt) != SQLITE_OK:
       dbError(db)
     if clear_bindings(stmtName.PStmt) != SQLITE_OK: